+++ src/NullMailer.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Mailer;
+
+/**
+ * This mailer can be used to avoid sending messages.
+ */
+final class NullMailer implements MailerInterface
+{
+    public function send(MessageInterface $message): void
+    {
+    }
+
+    public function sendMultiple(array $messages): SendResults
+    {
+        return new SendResults(array_values($messages), []);
+    }
+}